move delete logic to auth
